fix release note modal to fetch only when opened and not on mount
